Add support for new versions: - 1.125.2 - 1.126.1 - 1.127.0 - 1.128.3 - 1.131.3 - 1.132.1 - 1.133.0 - 1.134.8 - 1.134.9 - 1.134.10 - 1.135.0 - 1.138.0 - 1.139.5 - 1.139.6 - 1.140.0 - 1.142.0 - 1.148.1 - 1.148.2
Add supported games:
- Craft the World (thanks to a code fix in this commit, packets were wrongly addressed in the server side)
- Worms W.M.D

+++ Source/TestCases/Source/TestCase_ExtendedClient.cxx
@@ -59,7 +59,7 @@ void UpdateTimer() {
 						sender, (int)channel, std::string(buffer, out_msg_size)
 					));
 				std::string response = "SHOULDN'T HAPPEN CLIENT ACK: " + std::string(buffer, out_msg_size);
-				networking_ptr->SendP2PPacket(sender, response.c_str(), response.size(), P2P_SEND_UNRELIABLE, channel);
+				networking_ptr->SendP2PPacket(sender, response.c_str(), (uint32_t)response.size(), P2P_SEND_UNRELIABLE, channel);
 			}
 		}
 	}
